Set up router, moved Dashboard to ./pages
